# Use rust str instead of cstr in `ClassIdSource`