This simplifies things and improves performance of `ClassId::to_cow_str`, which is overhead in `CallContext::gd` and `check_rtti`.

```rust
#[bench(repeat = 10000)]
fn class_id_str() -> std::borrow::Cow<'static, str> {
    godot::classes::Object::class_id().to_cow_str()
}

// In debug build
// before:
//    -- class_id_str               ...      0.148μs      0.156μs
// after:
//    -- class_id_str               ...      0.119μs      0.125μs
```
I think the only reason for using cstr might be initializing StringName from it is slightly faster than rust str, but this is not a hot path since initialization only happens once.