@GodotRust 2025-09-24
API docs are being generated and will be shortly available at: https://godot-rust.github.io/docs/gdext/pr-1334